requirements.txt
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.0
django-safedelete==1.0.0
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
django-cors-headers==0.01